refactor controller routes and create movieFeatures.js

+++ controllers/commentController.js
@@ -10,7 +10,10 @@ exports.getComments = async (req, res, next) => {
       comments.reverse();
     }
 
-    return res.status(200).json(comments);
+    return res.status(200).json({
+      status: 200,
+      comments,
+    });
   } catch (error) {
     return next(new HttpError("Fetching comments failed", 500));
   }
@@ -19,31 +22,47 @@ exports.getComments = async (req, res, next) => {
 exports.getComment = async (req, res, next) => {
   try {
     const comment = await Comment.findByPk(req.params.commentId);
-    return res.status(200).json(comment);
+    return res.status(200).json({
+      status: 200,
+      comment,
+    });
   } catch (error) {
     return next(new HttpError("Fetching comment failed", 500));
   }
 };
 
 exports.createComment = async (req, res, next) => {
-  const ip = await publicIp.v4();
+  try {
+    let comment = req.body.comment.trim();
+    let movieId = +req.params.movieId;
+    let publicIp = await publicIp.v4();
 
-  const COMMENT_MODEL = {
-    comment: req.body.comment.trim(),
-    movieId: +req.params.movieId,
-    publicIp: ip,
-  };
-    if (req.body.comment.trim().length > 500) {
+    if (!comment || comment.length > 500) {
       return next(
         new HttpError(
-          "Comment length should not be more than 500 characters",
+          "Comment length should not be empty or more than 500 characters",
           500
         )
       );
     }
-  try {
-    const comment = await Comment.create(COMMENT_MODEL);
-    return res.status(200).json(comment);
+    if (!movieId || !publicIp) {
+      return next(
+        new HttpError(
+          "Please pass the movieId of the movie you want to comment on as a parameter on the route",
+          500
+        )
+      );
+    }
+    const COMMENT_MODEL = {
+      comment,
+      movieId,
+      publicIp,
+    };
+    const savedComment = await Comment.create(COMMENT_MODEL);
+    return res.status(200).json({
+      status: 200,
+      savedComment,
+    });
   } catch (error) {
     return next(new HttpError("Creating comment failed", 500));
   }
@@ -51,14 +70,38 @@ exports.createComment = async (req, res, next) => {
 
 exports.updateComment = async (req, res, next) => {
   try {
+    const comment = req.body.comment.trim();
+    const commentId = req.params.commentId;
+
+    if (!comment || comment.length > 500) {
+      return next(
+        new HttpError(
+          "Comment length should not be empty or more than 500 characters",
+          500
+        )
+      );
+    }
+
+    if (!commentId) {
+      return next(
+        new HttpError(
+          "Please pass the commentId of the comment you want to update as a parameter on the route",
+          500
+        )
+      );
+    }
+
     const COMMENT_MODEL = {
       comment: req.body.comment,
     };
 
-    const comment = await Comment.update(COMMENT_MODEL, {
-      where: { id: req.params.commentId },
+    const updatedComment = await Comment.update(COMMENT_MODEL, {
+      where: { id: commentId },
     });
-    return res.status(200).json(comment);
+    return res.status(200).json({
+      status: 200,
+      updatedComment,
+    });
   } catch (error) {
     return next(new HttpError("Updating comment failed", 500));
   }
@@ -66,6 +109,15 @@ exports.updateComment = async (req, res, next) => {
 
 exports.deleteComment = async (req, res, next) => {
   try {
+    const commentId = req.params.commentId;
+    if (!commentId) {
+      return next(
+        new HttpError(
+          "Please pass the commentId of the comment you want to update as a parameter on the route",
+          500
+        )
+      );
+    }
     const comment = await Comment.destroy({
       where: { id: req.params.commentId },
     });
